created form
created the form for the login information. This information consists of the  redirection  of the page onto another page which would be created the registration page.

The form itself consists of text fields to collect data and placeholders. These placeholders would be used in order to tell the user what type the data should be like.

The page was ended using a standard footer.